Added transactions to deleteOngoingDonation and added userID to OngoingDonation schema

+++ src/controllers/ongoingDonationsController.ts
@@ -1,5 +1,7 @@
+import mongoose from 'mongoose';
 import { Response, Request } from 'express';
 import { UploadedFile } from 'express-fileupload';
+import { User, UserDocument } from '../models/User/index';
 import { OngoingDonation, OngoingDonationDocument } from '../models/User/DonationForms';
 import { uploadImageAzure, deleteImageAzure } from '../util/azure-image';
 
@@ -79,7 +81,7 @@ export const updateOngoingDonation = (req: Request, res: Response) => {
  * @route DELETE /api/ongoingdonations/:donationID
  *
  */
-export const deleteOngoingDonation = (req: Request, res: Response) => {
+export const deleteOngoingDonation =  async (req: Request, res: Response) => {
     const donationId = req.params.donationID;
 
     if (!donationId) {
@@ -87,15 +89,56 @@ export const deleteOngoingDonation = (req: Request, res: Response) => {
         return;
     }
 
-    OngoingDonation.deleteOne({ _id: donationId })
-        .then((result) => {
-            if (result.deletedCount === 1) {
-                res.status(200).json({ message: 'Success' });
-            } else {
-                res.status(404).json({ message: 'The specified ongoing donation does not exist.' });
-            }
-        })
-        .catch((error: Error) => {
-            res.status(400).json({ message: error.message });
+    // Set up transaction session
+    const session = await mongoose.startSession();
+    session.startTransaction();
+
+    try {
+        // Get userId to modify status in User's donation array
+        const ongoingDonation:OngoingDonationDocument = await OngoingDonation.findById(donationId);
+        if (!ongoingDonation) {
+            throw new Error('Specified donation does not exist');
+        }
+        const userId = ongoingDonation.userID;
+
+        // Delete specified ongoing donation from queue
+        const deletedQueueResponse = await OngoingDonation.deleteOne({ _id: donationId }).session(session);
+        if (deletedQueueResponse.deletedCount !== 1) {
+            throw new Error('Could not delete donation from Ongoing Donation Queue.');
+        }
+
+        // Update status and ongoing of specified donation in User's donation array
+        const updatedDonationResponse:mongoose.UpdateWriteOpResult = await User.updateOne({ _id: userId, 'donations._id': donationId }, { $set: { 'donations.$.status': 'completed', 'donations.$.ongoing': false  } }).session(session);
+
+        if (updatedDonationResponse.nModified  !== 1) {
+            throw new Error('Could not update donation status for User.');
+        }
+
+        /*
+        OngoingDonation.deleteOne({ _id: donationId })
+            .then((result) => {
+                if (result.deletedCount === 1) {
+                    res.status(200).json({ message: 'Success' });
+                } else {
+                    res.status(404).json({ message: 'The specified ongoing donation does not exist.' });
+                }
+            })
+            .catch((error: Error) => {
+                res.status(400).json({ message: error.message });
+            }); */
+
+        // Commit transaction after deleting donation from queue and marking it as complete in array
+        await session.commitTransaction();
+        res.status(200).json({
+            message: 'Success',
+            donationForm: ongoingDonation
         });
+    } catch (err) {
+        await session.abortTransaction();
+        res.status(500).json({
+            message: err.message
+        });
+    } finally {
+        session.endSession();
+    }
 };
